requirements.txt
Pillow           
PyMuPDF          
pytesseract      
pygame           
gTTS             
pysimplegui      
tesseract        
tesseract-ocr
fitz
